Complete build fixes: warnings, submodules, and workflow improvements
Build Warning Fixes:
- Fix BLACKBONE_NO_TRACE macro redefinition with include guard
- Fix 32-bit to 64-bit shift conversion in MatrixCrypto.h (use 1LL)
- Fix size_t to int conversion warnings with static_cast in visual components
- Fix unreferenced variable warnings in Chat.cpp with (void)e suppression
- Fix double to float truncation in AmmoTracker.cpp

Workflow Improvements:
- Add explicit submodule initialization with authentication
- Fix ProtectMyTooling setup to use existing submodule instead of cloning
- Add diagnostic output for submodule status and directory contents
- Ensure ProtectMyTooling dependencies are installed when available

AmalgamLoader Submodule:
- Updated to latest commit with README improvements and legacy script removal

All build warnings resolved, ProtectMyTooling integration fixed, submodules properly initialized.

+++ Amalgam/src/Features/Visuals/AmmoTracker/AmmoTracker.cpp
@@ -299,7 +299,7 @@ void CAmmoTracker::DrawHudOverlay(const Vec3& vPlayerPos)
         const auto& [info, fTimeLeft] = vNearbySupplies[i];
         
         // Calculate position for this circle (offset left for multiple items)
-        int iCircleX = iStartX - (i * (iHudCircleSize * 2 + iHudSpacing)) + iHudCircleSize;
+        int iCircleX = iStartX - (static_cast<int>(i) * (iHudCircleSize * 2 + iHudSpacing)) + iHudCircleSize;
         int iCircleY = iStartY + iHudCircleSize;
         
         // Get colors
@@ -349,8 +349,8 @@ void CAmmoTracker::DrawClockFill(int iCenterX, int iCenterY, int iRadius, float 
     vertices.emplace_back(Vertex_t({ { static_cast<float>(iCenterX), static_cast<float>(iCenterY) } }));
     
     // Calculate pie slice
-    float flStartAngle = -90.0f * M_PI / 180.0f; // Start at top
-    float flEndAngle = flStartAngle + (2.0f * M_PI * flPercentage);
+    float flStartAngle = -90.0f * static_cast<float>(M_PI) / 180.0f; // Start at top
+    float flEndAngle = flStartAngle + (2.0f * static_cast<float>(M_PI) * flPercentage);
     float flStep = (flEndAngle - flStartAngle) / iVertices;
     
     // Add edge vertices
